backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==2.0.2
psycopg2-binary==2.9.9

# AI & Machine Learning
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production & Monitoring
gunicorn==21.2.0
sentry-sdk[fastapi]==1.38.0

# Caching (Optional)
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==23.2.0

# Health Checks
healthcheck==1.3.4
